requirements.txt
notion-client
mlx-lm
arxiv
requests
python-dotenv